pyproject.toml
[project]
name = "alixp-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "claude-agent-sdk>=0.1.6",
    "langsmith[claude-agent-sdk]>=0.4.40",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
    "mypy>=1.18.2",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
    "ruff>=0.14.3",
]
